backend/requirements.txt
django==5.2.8
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
environs==14.5.0
pandas==2.3.3
drf-spectacular==0.29.0
django-cors-headers==4.9.0